I made some pretty major changes to the catalog page. The actual contents are the same, but I changed the table layout to a layout of bootstrap panels. This makes scaling a lot better looking. I'm probably going to change some other things but this is how it's going to look pretty much.
I do plan on Adding a couple more types to the "special types" category, 
and I want to do something else besides the big panel asking for 
suggestions at the end, but I'm not sure what yet. I also want the order 
panel to scroll along with the page. I also want it to take the 
information from the catalog page and send it to the order page, but we 
may have to hold that off for our final...