This is Just to replace the pictures for the final design.
Right now I've got two ideas for additions. I'm planning on having the 
images on the catalog page clickable, and they'll show a drop-down that 
has the information and a way to add to a cart which will (hopefully) 
get sent to the order page. I'm also going to use some PHP to loop and 
create each catalog entry dynamically (grabbing from an SQL database).